# specification/swagger.yaml
---
swagger: "2.0"
info:
  description: "NIH Endotypes API"
  version: "1.0.0"
  title: "NIH Endotypes API"
  termsOfService: "None"
  contact:
    name: "Hong Yi"
    url: "https://github.com/renci/nih-endotypes-api"
    email: "hongyi@renci.org"
  license:
    name: "MIT"
    url: "http://opensource.org/licenses/MIT"
host: "localhost:7000"
basePath: "/v1"
schemes:
- "https"
paths:
  /endotypes:
    post:
      summary: "Get list of endotypes based on input"
      description: "Get list of endotypes based on input"
      operationId: "endotypes_get"
      produces:
        - "application/json"
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "input"
          required: true
          schema:
            $ref: "#/definitions/input"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/inline_response_200"
        400:
          description: "Invalid parameter"
        404:
          description: "No endotypes found"
        500:
          description: "Server error"
      x-swagger-router-controller: "controllers.default_controller"
definitions:
  exposure:
    properties:
      value:
        type: "string"
      units:
        type: "string"
      exposure_type:
        type: "string"
  visit:
    properties:
      time:
        type: "string"
      zip:
        type: "string"
      visit_type:
        type: "string"
      icd_code:
        type: "string"
      exposure:
        $ref: "#/definitions/exposure"
  input:
    properties:
      date_of_birth:
        type: "string"
      race:
        type: "string"
      sex:
        type: "string"
      visits:
        type: "array"
        items:
          $ref: "#/definitions/visit"
  endotypes_period:
    properties:
      start_time:
        type: "string"
        format: "date-time"
      end_time:
        type: "string"
        format: "date-time"
  inline_response_200:
    properties:
      endotype_id:
        type: "string"
      endotype_description:
        type: "string"
      periods:
        type: "array"
        items:
          $ref: "#/definitions/endotypes_period"